+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/message/FileBasedRoomMessageTimelineElementViewModel.kt
@@ -36,8 +36,14 @@ abstract class FileBasedRoomMessageTimelineElementViewModel<C : RoomMessageEvent
 
     private val _loadMediaResultPlatformMedia: MutableStateFlow<PlatformMedia?> = MutableStateFlow(null)
     override val loadMediaResultPlatformMedia: StateFlow<PlatformMedia?> = _loadMediaResultPlatformMedia.asStateFlow()
-    private val _loadMediaResult: MutableStateFlow<ByteArray?> = MutableStateFlow(null)
-    override val loadMediaResult: StateFlow<ByteArray?> = _loadMediaResult.asStateFlow()
+    private val _loadMediaResultBytes: MutableStateFlow<ByteArray?> = MutableStateFlow(null)
+    override val loadMediaResultBytes: StateFlow<ByteArray?> = _loadMediaResultBytes.asStateFlow()
+
+    @Deprecated(
+        "This will be removed in the future for consistency with downloadMedia behaviour, please use loadMediaResultBytes instead",
+        replaceWith = ReplaceWith("loadMediaResultBytes")
+    )
+    override val loadMediaResult: StateFlow<ByteArray?> = loadMediaResultBytes
     private val _loadMediaProgress: MutableStateFlow<FileTransferProgressElement?> = MutableStateFlow(null)
     override val loadMediaProgress: StateFlow<FileTransferProgressElement?> = _loadMediaProgress.asStateFlow()
     private val _loadMediaError: MutableStateFlow<String?> = MutableStateFlow(null)
@@ -79,7 +85,7 @@ abstract class FileBasedRoomMessageTimelineElementViewModel<C : RoomMessageEvent
             resultAsync.await()
                 .onSuccess {
                     _loadMediaResultPlatformMedia.value = it
-                    _loadMediaResult.value = it.toByteArray(
+                    _loadMediaResultBytes.value = it.toByteArray(
                         coroutineScope,
                         expectedSize = content.info?.size,
                         maxSize = maxMediaSizeInMemory