+++ CHANGELOG.md
@@ -12,14 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Technical: Support deep nested settings view
 
 ### Changed
-- Expose platform media in loadMedia of FileBasedRoomMessageTimelineElementViewModel to allow temporary file generation
 
 - iOS: startMessenger() helper function now also returns MatrixMultiMessenger in order to bridge certain values to Swift
   code (like objects that can only be obtained via Koin)
 - Don't close DI, because garbage collector can handle it
 - Upgrade Trixnity to 4.22.3
+- Expose platform media in loadMedia of FileBasedRoomMessageTimelineElementViewModel to allow temporary file generation
 
 ### Deprecated
+
+- rename loadMediaResult to loadMediaResultBytes in FileBased RoomMessageEventContent 
 
 ### Removed